web: structure: compute card bounds row-by-row
concourse/concourse#5921

In order to support dynamically sized rows due to all cards in a row
having a common header height, group cards by row and handle each row
individually (for computing the grid coordinates).

This is technically a minor behaviour change as visibility is now based
on pixel coordinates rather than grid coordinates, and this fixes a bug
around the team header not being used to offset the pipeline cards
properly. I think it's more valuable to view this as a structural
change, though - it simplifies the grid coordinate logic and removes
some duplication between the favorites section and the all pipelines
section.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>