web: behaviour: pipeline cards display instance vars
concourse/concourse#5921

On the instance group view, we now display the instance vars using the
flattened notation (i.e. {a: {b: 1, c: 2}} -> a.b: 1, a.c: 2) - this is
also used by the across step subheaders) in the pipeline card header.

All cards in a row have the same header height, regardless of whether
the cards need that much space (we take the maximum of all cards in the
row).  Typically, cards in an instance group will have the same number
of instance vars - this is to handle the odd cases.

The cards themselves expand along with the header height, meaning that
the rows are dynamically sized.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>
Co-authored-by: Aidan Oldershaw <aoldershaw@pivotal.io>