web: behaviour: add pipeline preview tooltip
concourse/concourse#5921

Uses the hyphen-separated values notation, at least for now. I also
noticed that when you hover over a tooltip at a point that intersects
the element (e.g. on the arrow), you unhover the element, so the tooltip
disappears - and then you're hovering again, so it reappears - and this
repeats really quickly, producing a flickering effect. So, I added
pointer-events: none to the tooltip so it doesn't trigger mouse events.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>